+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.62.0] - 2022-06-29
+
 ### Added
 - `count` to `productSuggestions` query.